+++ src/core/ports/data-management.repository.ts
@@ -1,0 +1,6 @@
+import { Project } from '../entities/Project';
+
+export interface DataManagementRepository {
+  getAll(): Promise<string>;
+  replaceAll(projects: Project[]): Promise<void>;
+}